requirements.txt
slack-bolt==1.18.0
openai==1.3.0
matplotlib==3.7.2
pandas==2.1.0
numpy==1.24.3
requests==2.31.0
gunicorn==21.2.0
flask==2.3.3
python-dotenv==1.0.0
httpx>=0.24.0,<0.28.0